change to IntConstMult to avoid limitations, enable calculation up to wIn=63 bit.

+++ src/BitHeap/ModuloBarrett.hpp
@@ -13,7 +13,7 @@ namespace flopoco {
 
         int wIn;
         int modulo;
-        int m;
+        long long int m;
         int k;
